Extract renderer-agnostic buffer classes from WebGL layer (#1410)
- Move VertexArrayBuffer from video/webgl/buffer/ to video/buffer/
  (zero GL dependency — pure typed array management)
- Split IndexBuffer into renderer-agnostic IndexBuffer base class
  (data accumulation, quad patterns) and WebGLIndexBuffer subclass
  (GL buffer creation, bind, upload)
- Update all imports in batchers and tests

Part of #1410 (TextureCache + Batcher abstraction). The TextureCache
unit management refactoring is deferred pending investigation of a
rendering regression with per-sprite shader effects.
