+++ packages/melonjs/tests/vertexBuffer.spec.js
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import VertexArrayBuffer from "../src/video/webgl/buffer/vertex.js";
+import VertexArrayBuffer from "../src/video/buffer/vertex.js";
 import { buildMultiTextureFragment } from "../src/video/webgl/shaders/multitexture.js";
 
 describe("VertexArrayBuffer", () => {
